Assimp: Split of files, and reworked skeletons loading.
